requirements.txt
fastapi
uvicorn[standard]
pydantic
pytest
requests  # Useful for more complex testing scenarios.
bcrypt
python-dotenv
sqlalchemy